requirements.txt
flask>=2.2
redis>=4.5
python-dotenv>=1.0
